move AIMET encoding parsing utils into a separate file
